+++ libraries/session/src/main/java/androidx/media3/session/ConnectionState.java
@@ -60,12 +60,15 @@ import java.util.List;
 
   @Nullable public final Token platformToken;
 
+  public final ImmutableList<CommandButton> commandButtonsForMediaItems;
+
   public ConnectionState(
       int libraryVersion,
       int sessionInterfaceVersion,
       IMediaSession sessionBinder,
       @Nullable PendingIntent sessionActivity,
       ImmutableList<CommandButton> customLayout,
+      ImmutableList<CommandButton> commandButtonsForMediaItems,
       SessionCommands sessionCommands,
       Player.Commands playerCommandsFromSession,
       Player.Commands playerCommandsFromPlayer,
@@ -78,6 +81,7 @@ import java.util.List;
     this.sessionBinder = sessionBinder;
     this.sessionActivity = sessionActivity;
     this.customLayout = customLayout;
+    this.commandButtonsForMediaItems = commandButtonsForMediaItems;
     this.sessionCommands = sessionCommands;
     this.playerCommandsFromSession = playerCommandsFromSession;
     this.playerCommandsFromPlayer = playerCommandsFromPlayer;
@@ -91,6 +95,7 @@ import java.util.List;
   private static final String FIELD_SESSION_BINDER = Util.intToStringMaxRadix(1);
   private static final String FIELD_SESSION_ACTIVITY = Util.intToStringMaxRadix(2);
   private static final String FIELD_CUSTOM_LAYOUT = Util.intToStringMaxRadix(9);
+  private static final String FIELD_COMMAND_BUTTONS_FOR_MEDIA_ITEMS = Util.intToStringMaxRadix(13);
   private static final String FIELD_SESSION_COMMANDS = Util.intToStringMaxRadix(3);
   private static final String FIELD_PLAYER_COMMANDS_FROM_SESSION = Util.intToStringMaxRadix(4);
   private static final String FIELD_PLAYER_COMMANDS_FROM_PLAYER = Util.intToStringMaxRadix(5);
@@ -101,7 +106,7 @@ import java.util.List;
   private static final String FIELD_IN_PROCESS_BINDER = Util.intToStringMaxRadix(10);
   private static final String FIELD_PLATFORM_TOKEN = Util.intToStringMaxRadix(12);
 
-  // Next field key = 13
+  // Next field key = 14
 
   public Bundle toBundleForRemoteProcess(int controllerInterfaceVersion) {
     Bundle bundle = new Bundle();
@@ -112,6 +117,12 @@ import java.util.List;
       bundle.putParcelableArrayList(
           FIELD_CUSTOM_LAYOUT,
           BundleCollectionUtil.toBundleArrayList(customLayout, CommandButton::toBundle));
+    }
+    if (!commandButtonsForMediaItems.isEmpty()) {
+      bundle.putParcelableArrayList(
+          FIELD_COMMAND_BUTTONS_FOR_MEDIA_ITEMS,
+          BundleCollectionUtil.toBundleArrayList(
+              commandButtonsForMediaItems, CommandButton::toBundle));
     }
     bundle.putBundle(FIELD_SESSION_COMMANDS, sessionCommands.toBundle());
     bundle.putBundle(FIELD_PLAYER_COMMANDS_FROM_SESSION, playerCommandsFromSession.toBundle());
@@ -161,6 +172,15 @@ import java.util.List;
             ? BundleCollectionUtil.fromBundleList(
                 b -> CommandButton.fromBundle(b, sessionInterfaceVersion), commandButtonArrayList)
             : ImmutableList.of();
+    @Nullable
+    List<Bundle> commandButtonsForMediaItemsArrayList =
+        bundle.getParcelableArrayList(FIELD_COMMAND_BUTTONS_FOR_MEDIA_ITEMS);
+    ImmutableList<CommandButton> commandButtonsForMediaItems =
+        commandButtonsForMediaItemsArrayList != null
+            ? BundleCollectionUtil.fromBundleList(
+                b -> CommandButton.fromBundle(b, sessionInterfaceVersion),
+                commandButtonsForMediaItemsArrayList)
+            : ImmutableList.of();
     @Nullable Bundle sessionCommandsBundle = bundle.getBundle(FIELD_SESSION_COMMANDS);
     SessionCommands sessionCommands =
         sessionCommandsBundle == null
@@ -192,6 +212,7 @@ import java.util.List;
         IMediaSession.Stub.asInterface(sessionBinder),
         sessionActivity,
         customLayout,
+        commandButtonsForMediaItems,
         sessionCommands,
         playerCommandsFromSession,
         playerCommandsFromPlayer,