+++ libraries/session/src/main/java/androidx/media3/session/MediaSessionStub.java
@@ -535,6 +535,7 @@ import java.util.concurrent.ExecutionException;
                     connectionResult.customLayout != null
                         ? connectionResult.customLayout
                         : sessionImpl.getCustomLayout(),
+                    sessionImpl.getCommandButtonsForMediaItems(),
                     connectionResult.availableSessionCommands,
                     connectionResult.availablePlayerCommands,
                     playerWrapper.getAvailableCommands(),